+++ android/history.md
@@ -3,6 +3,9 @@
 ## 13.0 alpha
 * Start version 13.0
 
+## 2019-08-05 12.0.4075 beta
+* Adjustments to Settings UI
+
 ## 2019-07-29 12.0.4074 beta
 * Initial beta release of Keyman for Android 12
 * [Pull Requests](https://github.com/keymanapp/keyman/pulls?utf8=%E2%9C%93&q=is%3Apr+merged%3A2019-02-25..2019-08-04+label%3Aandroid+base%3Amaster)